Add compressed copies dated January 27, 2020
- Add compressed copies of the Virtual-Universe-Early-Dev repository as of January 27, 2020 in preparation for the major commit coming later this week.

NOTE: This copy does not currently build.  We may provide a fixed version of this shortly.  However, we advise we may not do much more with this version other then providing a fixed build.  We are moving to the next version of our architecture.  If you plan to use our open source architecture to run your grid, we advise you to stay up to date with the officially supported version we provide.
Signed-off-by: NoahStarfinder <noah@secondgalaxy.com>